events: make path hash more resilient
By replacing the uint32 hash with a double, the same code works for the
othello game. Also add code that does a full path comparison (should
profile/assess if this is better than storing full paths and taking a hit
during event processing instead.)

+++ lib/mrg-internal.h
@@ -92,7 +92,7 @@ typedef struct MrgItem {
   float          y1;
 
   cairo_path_t   *path;
-  uint32_t        path_hash;
+  double          path_hash;
 
   MrgType   types; /* all cb's ored together */
   MrgItemCb cb[MRG_MAX_CBS];
